+++ apps/user/src/components/common/NeedLoginModal/NeedLoginModal.tsx
@@ -12,12 +12,15 @@ const NeedLoginModal = ({ isOpen, onClose, onConfirm }: NeedLoginModalProps) => 
   useEffect(() => {
     if (isOpen) {
       document.body.style.overflow = 'hidden';
+      document.documentElement.style.overflow = 'hidden';
     } else {
-      document.body.style.overflow = 'auto';
+      document.body.style.overflow = '';
+      document.documentElement.style.overflow = '';
     }
 
     return () => {
-      document.body.style.overflow = 'auto';
+      document.body.style.overflow = '';
+      document.documentElement.style.overflow = '';
     };
   }, [isOpen]);
